relase version

+++ mobileSafe/src/main/java/com/jcxy/MobileSafe/engine/TaskInfos.java
@@ -7,8 +7,6 @@ import android.content.pm.PackageInfo;
 import android.content.pm.PackageManager;
 import android.graphics.drawable.Drawable;
 import android.os.Debug;
-import android.text.format.Formatter;
-import android.widget.Toast;
 
 import com.jcxy.MobileSafe.bean.TaskInfo;
 import com.wenming.library.processutil.ProcessManager;
@@ -23,9 +21,9 @@ import java.util.List;
 public class TaskInfos {
     private Context context;
     private ActivityManager manager;
-    private PackageManager pm;
+    private  PackageManager pm;
     private TaskInfo info;
-    List<TaskInfo> taskInfoList ;
+    List<TaskInfo> taskInfoList;
     private List<ActivityManager.RunningAppProcessInfo> list;
 
     public TaskInfos(Context context) {
@@ -41,7 +39,7 @@ public class TaskInfos {
 
     public List<TaskInfo> getTasks() {
 
-        list =ProcessManager.getRunningAppProcessInfo(context);
+        list = ProcessManager.getRunningAppProcessInfo(context);
 
         if (list != null && list.size() > 0) {
             taskInfoList = new ArrayList<TaskInfo>();
@@ -112,22 +110,25 @@ public class TaskInfos {
         return 0;
     }
 
-    public  void killAllProcess(){
+    private int memeory = 0;
+    private int count = 0;
+
+    public String killAllProcess() {
         List<TaskInfo> tasks = getTasks();
-        int memeory=0;
-        int count=0;
-        ActivityManager manager= (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
-        for ( TaskInfo info: tasks) {
-            String packName = info.getPackName();
-            if(context.getPackageName().equals(packName)){
-                continue;
+        ActivityManager manager = (ActivityManager) context.getSystemService(Context.ACTIVITY_SERVICE);
+        for (TaskInfo info : tasks) {
+            if (!info.getSystemProcess()) {
+                String packName = info.getPackName();
+                if (context.getPackageName().equals(packName)) {
+                    continue;
+                }
+                memeory += info.getProcessSize();
+                count++;
+                manager.killBackgroundProcesses(packName);
             }
-            memeory+=info.getProcessSize();
-            count++;
-            manager.killBackgroundProcesses(packName);
+
         }
-        Toast.makeText(context,"共清理了"+count+"个进程,释放内存"+ Formatter.formatFileSize(context,memeory),Toast.LENGTH_SHORT).show();
-        manager=null;
-        tasks=null;
+        return memeory+":"+count;
+
     }
 }
